+++ lib/src/number/number.dart
@@ -20,7 +20,7 @@ export 'number/real.dart';
 export 'large_numbers/large.dart';
 
 export 'util/converters.dart';
-export 'util/erf.dart';
+
 export 'util/jenkins_hash.dart';
 export 'util/romans_exception.dart';