+++ lib/src/number/util/utils.dart
@@ -1,3 +1,2 @@
-export 'erf.dart';
 export 'jenkins_hash.dart';
 export 'romans_exception.dart';